Refactor to re-organize directory structure
- Move React components to `src/components/`
- Move React hooks to `src/hooks/`
- Remove `utils/index`
