+++ src/pages/Blog.jsx
@@ -1,6 +1,8 @@
-import { Link } from "../Link";
 import { blogs } from "../blog/index";
-import { Breadcrumb, Time, useTitle } from "../utils";
+import { Breadcrumb } from "../components/Breadcrumb";
+import { Link } from "../components/Link";
+import { Time } from "../components/Time";
+import { useTitle } from "../hooks/useTitle";
 
 export const Blog = () => {
   useTitle("Blog");
