--- src/utils/index.js
@@ -1,7 +1,0 @@
-export * from "./Breadcrumb.jsx";
-export * from "./CopyToClipboard.jsx";
-export * from "./Time.jsx";
-
-export * from "./assert.js";
-export * from "./useMount.js";
-export * from "./useTitle.js";
